allow HTTP version at the end of an HTTP 0.9 request line
it is a bit of a gray area, but  we believe a fully compliant server should accept 0.9 as a version. 

According to https://www.w3.org/Protocols/HTTP/Request.html 
> If the protocol version is not specified, the server assumes that the browser uses HTTP version 0.9
this sets 0.9 to the default, but doesn't say that it should not be set. 

From https://tools.ietf.org/html/rfc1945#section-5.1 also 
> This document defines both the 0.9 and 1.0 versions of the HTTP protocol.
and 
> The Request-Line begins with a method token, followed by the
>   Request-URI and the protocol version, and ending with CRLF. The
>   elements are separated by SP characters. No CR or LF are allowed
>   except in the final CRLF sequence.
>
>       Request-Line = Method SP Request-URI SP HTTP-Version CRLF